Add 104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,7 +9,8 @@
 int main(void)
 {
 	int i;
-	long double x = 1, y = 2, f;
+
+	unsigned long x = 1, y = 2, f;
 
 	for (i = 1; i <= 98; i++)
 	{
@@ -20,7 +21,7 @@ int main(void)
 			f = x + y;
 			x = y;
 			y = f;
-			printf("%.0Lf", f);
+			printf("%.0lu", f);
 		}
 
 		if (i != 98)
